IfiS2021/IN5400/Assignment1/requirements.txt
numpy
torch
sklearn
pillow
flask
tqdm
pandas
torchvision
